ksh93/sh.1: say that ${.sh.subshell} only counts virtual subshells
${.sh.subshell} only counts the nesting level of virtual (non-forked)
subshell environments relative to the current process, whether that
is the main shell environment or a real (forked) subshell.

This means ${.sh.subshell} is not sufficient to test if your code
is running in a subshell environment, so it's important to mention
this in the manual.

E.g.:
    (echo ${.sh.subshell}; ulimit -t 1; echo ${.sh.subshell}); :
outputs 1 followed by 0.

(cherry picked from commit 63dad8863fd26bf3aa4d7a1cf60b743f4169f7c4)