job_init(): fix init on setpgid() permission denied (re: 41ebb55a)
Symptoms of this bug below. These only seem to occur on Linux and
only if you replace your initial login shell by ksh using 'exec'.

1. An erroneous 'Interrupt' message is printed after stopping the
   read builtin in a script. Reproducer:

	$ exec arch/*/bin/ksh
	$ cat ./reproducer.sh
	#!/bin/sh
	read foo
	$ ./reproducer.sh
	^C$ <Enter>
	[1] + Interrupt                ../reproducer.sh

2. Ctrl+C fails to stop /bin/package make. Reproducer:

	$ exec arch/*/bin/ksh
	$ mv arch arch.old
	$ bin/package make
	# Press Ctrl+C multiple times

Analysis: In 41ebb55a, I made an error in changing job_init() to
work correctly on non-interactive shells. This line from before:

552|	if(possible = (setpgid(0,job.mypgid)>=0) || errno==EPERM)

was changed to:

555|	possible = (setpgid(0,job.mypgid) >= 0);
556|	if(sh_isoption(SH_INTERACTIVE) && (possible || errno==EPERM))

That is wrong. Before, 'possible' was set to 1 (true) if setpgid()
either succeeded or failed with EPERM. After, it is only set to 1
if setpgid() succeeds. As a result, job control initialisation is
aborted later on upon a test for non-zero 'possible'.

src/cmd/ksh93/sh/jobs.c: job_init():
- Once again set possible to 1 even if setpgid() fails with EPERM.

Thanks to @JohnoKing for the bug report and reproducers.

Resolves: https://github.com/ksh93/ksh/issues/210